favorites

+++ bin/wallpaper.sh
@@ -33,7 +33,12 @@ then
     title=$(cat /tmp/wallpaper.json | jq .title | tr -d '"');    
     # give a browser user agent to work around pixabay.
     # The download is initiated by a me from the keyboard
-    wget --header='User-Agent: Mozilla/5.0 (Windows NT 6.1; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/64.0.3282.167 Safari/537.36' --no-http-keep-alive -d -O /tmp/wallpaper.jpg $imageLink
+    if [ -f ${imageLink} ]        
+    then
+        cp ${imageLink} /tmp/wallpaper.jpg
+    else
+        wget --header='User-Agent: Mozilla/5.0 (Windows NT 6.1; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/64.0.3282.167 Safari/537.36' --no-http-keep-alive -d -O /tmp/wallpaper.jpg $imageLink
+    fi
     # Use convert to inject provider details (Required by Pixabay's high resolution requirements)
     convert -background black -pointsize 16 -fill white -annotate +10+50 "[${provider}] ${title}" /tmp/wallpaper.jpg /tmp/convert.jpg
     #eog /tmp/convert.jpg
